+++ tests/pest/Feature/VariableTest.php
@@ -12,7 +12,6 @@ use putyourlightson\sprig\variables\SprigVariable;
 beforeEach(function() {
     Sprig::bootstrap();
     Craft::$app->getResponse()->getHeaders()->removeAll();
-    Craft::$app->getView()->clear();
     Craft::$app->getView()->setTemplatesPath(Craft::getAlias('@putyourlightson/sprig/test/templates'));
 });
 
@@ -50,7 +49,7 @@ describe('Sprig request', function() {
             'sprig:config' => Craft::$app->getSecurity()->hashData('{"template":"test"}'),
         ]);
         Craft::$app->getRequest()->getHeaders()->set('HX-Request', true);
-        Craft::$app->getView()->clear();
+        Sprig::$core->requests->resetOobSwapSources();
     });
 
     test('Swap OOB', function() {